[ota-requestor] Move interfaces to include/platform
