+++ examples/ota-requestor-app/linux/LinuxOTARequestorDriver.h
@@ -19,7 +19,7 @@
 /* This file contains the decalarions for the Linux implementation of the
  * the OTARequestorDriver interface class
  */
-#include "app/clusters/ota-requestor/OTARequestorDriver.h"
+#include <platform/OTARequestorDriver.h>
 
 namespace chip {
