+++ src/app/clusters/ota-requestor/ClusterInterface.cpp
@@ -20,7 +20,7 @@
  * to the OTA Requestor object that handles them
  */
 
-#include "OTARequestorInterface.h"
+#include <platform/OTARequestorInterface.h>
 
 // OTA Software Update Requestor Cluster AnnounceOtaProvider Command callback
 bool emberAfOtaSoftwareUpdateRequestorClusterAnnounceOtaProviderCallback(
