+++ src/app/clusters/ota-requestor/OTARequestor.h
@@ -24,11 +24,11 @@
 
 #include <app/CASESessionManager.h>
 #include <app/server/Server.h>
+#include <platform/OTARequestorDriver.h>
+#include <platform/OTARequestorInterface.h>
 #include <protocols/bdx/BdxMessages.h>
 
 #include "BDXDownloader.h"
-#include "OTARequestorDriver.h"
-#include "OTARequestorInterface.h"
 
 namespace chip {
