+++ src/include/platform/OTAImageProcessor.h
